+++ src/plugins/piral-pwa/src/index.ts
@@ -1,3 +1,0 @@
-export * from './create';
-export * from './setup';
-export * from './types';