+++ src/utilities/piral-systemjs-utils/vendor.js
@@ -1,3 +1,0 @@
-require('systemjs/dist/extras/amd');
-require('systemjs/dist/extras/named-exports');
-require('systemjs/dist/extras/use-default');