+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 1.0.0 (tbd)
 
 - First LTS release
-- Removed `piral-native` package
-- Removed `piral-ssr-utils` package
+- Removed `piral-native` framework package
+- Removed `piral-ssr-utils` utils package
+- Removed `piral-systemjs-utils` utils package
+- Removed `piral-pwa` plugin package
 
 ## 0.15.13 (May 31, 2023)